Extract player from main (but not well)
This commit puts Player into its own file, but ends
up tightly coupling screen dimensions to that it,
which is not sustainable long-term.
